bears: Move results into coalib root src
Results are a core part of coala. They deserve a place in our root
source directory especially since there will be ResultAction coming
soon!
