+++ coalib/results/LineResult.py
@@ -1,5 +1,5 @@
-from coalib.bears.results.RESULT_SEVERITY import RESULT_SEVERITY
-from coalib.bears.results.Result import Result
+from coalib.results.RESULT_SEVERITY import RESULT_SEVERITY
+from coalib.results.Result import Result
 
 
 class LineResult(Result):
