+++ coalib/results/RESULT_SEVERITY.py
