+++ setup.py
@@ -25,7 +25,6 @@ if __name__ == "__main__":
                     'coalib.bearlib.abstractions',
                     'coalib.bearlib.spacing',
                     'coalib.bears',
-                    'coalib.bears.results',
                     'coalib.collecting',
                     'coalib.misc',
                     'coalib.output',
@@ -33,6 +32,7 @@ if __name__ == "__main__":
                     'coalib.parsing',
                     'coalib.processes',
                     'coalib.processes.communication',
+                    'coalib.results',
                     'coalib.settings',
           ],
           license="AGPL v3",
